requirements.txt
Ghost.py
Flask
boto
cloudinary